% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{add_local_density}
\alias{add_local_density}
\title{Add local density}
\usage{
add_local_density(.data, x, y, var = "local_density", sort_output = TRUE)
}
\arguments{
\item{.data}{input data frame}

\item{x}{unquoted column name of variable to use for the x-variable}

\item{y}{unquoted column name of variable to use for the y-variable}

\item{var}{name to use for the output column}

\item{sort_output}{if TRUE, return the output sorted by local density}
}
\description{
Add a column of local density to a data frame
}
\details{
The \code{sort_output} parameter sometimes helps the output plot
look a little less speckled.
}
