% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{args2ge}
\alias{args2ge}
\title{Assign the default argument values of a function in the global environment}
\usage{
args2ge(f)
}
\arguments{
\item{f}{a function}
}
\description{
Assign the default argument values of a function in the global environment
}
