% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broad_colors}
\alias{broad_colors}
\title{Broad colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
broad_colors
}
\description{
A vector of colors from the Broad Style guide.
}
\examples{
\dontrun{p + scale_color_manual(values = broad_colors)}
}
\keyword{datasets}
