% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{corner}
\alias{corner}
\title{Show the top-left corner}
\usage{
corner(x, nr = 5, nc = 5)
}
\arguments{
\item{x}{the input data frame or matrix}

\item{nr}{number of rows to pull out}

\item{nc}{number of columns to pull out}
}
\description{
This function pulls out the top-left corner of a data frame or matrix. This
is useful if the input is very tall and/or wide, such that printing the
entire object fills up the entire console, even with the nice printing
features of e.g. tibbles.
}
\details{
use negative nr or nc to select bottom or right corners, respectively.
}
