% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{find_dups}
\alias{find_dups}
\title{Find duplicates}
\usage{
find_dups(data, var, order_by = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{var}{a variable in that data frame (unquoted)}

\item{order_by}{if TRUE, order the result by the specified variable}
}
\description{
Filter to rows where the specified variable has at least one dulicate.
Useful for identifying problematic cases when 1:1 joins come out at 1:many.
}
\examples{
mtcars |> find_dups(mpg)
}
