% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{fpat}
\alias{fpat}
\title{Filter to rows matching a pattern}
\usage{
fpat(x, pattern, col_name)
}
\arguments{
\item{x}{input data frame}

\item{pattern}{pattern to search for}

\item{col_name}{column to search}
}
\description{
Filter to rows matching a pattern
}
