% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_local_density}
\alias{get_local_density}
\title{Get local 2D density}
\usage{
get_local_density(x, y, group = NULL, ...)
}
\arguments{
\item{x}{a vector of x-values}

\item{y}{a vector of y-values}

\item{group}{a group}

\item{...}{arguments to pass to ks::kde for the density estimation}
}
\description{
Compute a 2D density estimate and return the estimate evaluated
at each pair of points
}
\details{
This is useful for computing color-density scatterplots.
}
\examples{
overlapping = data.frame(x = rt(3e3, df = 3), y = rt(3e3, df = 3))
get_local_density(overlapping$x, overlapping$y)
}
