% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{has_dups}
\alias{has_dups}
\title{Has duplicates}
\usage{
has_dups(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical vector the same length as the input
}
\description{
Return TRUE on all elements of a vector that has any duplicates (i.e. including the first occurrence)
}
\seealso{
\code{\link[vctrs:vec_duplicate]{vctrs::vec_duplicate_detect()}} (will be faster for large x)
}
