% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{pull1}
\alias{pull1}
\title{Pull out a single element from a (list-) variable}
\usage{
pull1(.data, my_var, one_to_pull = 1)
}
\arguments{
\item{.data}{a data frame}

\item{my_var}{the unquoted name of the variable you want to pull out an element from}

\item{one_to_pull}{if you don't want to pull the first element, set this to the row number of the one you want}
}
\value{
The \code{one_to_pull}-th element of the input variable
}
\description{
\code{dplyr::pull} is a convenient way to pull out the values of a column
from a data frame, but if you've got a long list column and just want to
check one to make sure everything's working, printing the whole list can
clutter up the console. \code{pull1} just pulls out one.
}
