% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{show_classes}
\alias{show_classes}
\title{Show the classes of a data frame}
\usage{
show_classes(x)
}
\arguments{
\item{x}{input data frame}
}
\description{
Return a data frame giving the class of each column of the input
}
