% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary_tbl}
\alias{summary_tbl}
\title{Summarise a stanfit in a tibble}
\usage{
summary_tbl(stan_fit, ...)
}
\arguments{
\item{stan_fit}{a stanfit object to summarise}

\item{...}{arguments passed to \verb{summary,stanfit-method} e.g. pars or probs}
}
\value{
a summary tibble
}
\description{
Get a stan summary as a tibble rather than a list of matrices
}
\seealso{
\code{\link[rstan:summary-methods]{summary}}
}
