% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{theme_pres}
\alias{theme_pres}
\title{A theme for presentations}
\usage{
theme_pres(base_size = 22, ...)
}
\arguments{
\item{base_size}{The baseline size of plot elements}

\item{...}{arguments to pass to ggplot2::theme()}
}
\description{
Ever seen oblivious researchers presenting slides with ggplot2's
characteristic tiny axis-titles that even Legolas's elf eyes couldn't see?
This theme makes text larger so that the output plots look nice on slides.
Otherwise it's mostly identical to \code{ggplot2::theme_gray()}.
}
