% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{timer}
\alias{timer}
\title{Set a timer.}
\usage{
timer(m = 5, sound = 9)
}
\arguments{
\item{m}{time in minutes}

\item{sound}{beepr::beep() sound number}
}
\description{
"wilhelm" is the best sound, you know it.
}
\examples{
timer(2/60)
}
