% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dep_graph.R
\name{uniq_pkg_deps}
\alias{uniq_pkg_deps}
\title{Count unique dependencies}
\usage{
uniq_pkg_deps(pkg, order = 1)
}
\arguments{
\item{pkg}{a package to check}

\item{order}{Consider combinations of this many packages. Be careful going
beyond 3.}
}
\value{
A data.table listing the packages and the number of unique
dependencies.
}
\description{
A function for package developers to count which direct
dependencies add the most unique dependencies (direct plus downstream) in
total.
}
\details{
Use this function to identify which packages have the most unique
downstream dependencies. Set the order argument to check which pairs,
triplets, etc have the most unique dependencies.
}
