% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{ws_size}
\alias{ws_size}
\title{See the memory usage of objects in your workspace

It's easy to see the memory usage of a single object with
\code{object.size()} but doing that in a sorted, pretty way for everything in
the workspace is a bit more involved. This function has taken care of that.}
\usage{
ws_size()
}
\value{
A tibble of objects in the global environment and columns stating
how much memory the objects are using (both with pretty units and the raw
integer number of bytes).
}
\description{
See the memory usage of objects in your workspace

It's easy to see the memory usage of a single object with
\code{object.size()} but doing that in a sorted, pretty way for everything in
the workspace is a bit more involved. This function has taken care of that.
}
